hf_face_api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
numpy==1.26.4
insightface==0.7.3
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
Pillow==10.2.0
pydantic==2.6.1
